chore: refactoring and folder restructuring
