+++ examples-standalone/grid-charts-integration/src/app/components/action-buttons/action-buttons.component.ts
@@ -1,9 +1,9 @@
 import { Component } from '@angular/core';
 
 @Component({
-  selector: 'app-action-buttons',
-  templateUrl: './action-buttons.component.html',
-  styleUrls: ['./action-buttons.component.scss']
+    selector: 'app-action-buttons',
+    templateUrl: './action-buttons.component.html',
+    styleUrls: ['./action-buttons.component.scss']
 })
 export class ActionButtonsComponent {
