+++ examples-standalone/grid-charts-integration/src/app/components/common/select-chart-type-btngroup/select-chart-type-btngroup.component.html
@@ -1,0 +1,10 @@
+<label>Selected chart type:
+    <kendo-buttongroup [selection]="'single'">
+        <button #button *ngFor="let chart of data" kendoButton
+            [toggleable]="true"
+            [selected]="isSelected(chart)"
+            (selectedChange)="selectedChange($event, chart)">
+            {{chart}}
+        </button>
+    </kendo-buttongroup>
+</label>
