+++ examples-standalone/grid-charts-integration/src/app/components/common/select-chart-type/select-chart-type.component.html
@@ -1,0 +1,9 @@
+<label>Selected chart type:
+	<kendo-dropdownlist
+		style="width: 100%"
+		[data]="data"
+		[value]="chartName"
+		(valueChange)="onChange($event)"
+	>
+	</kendo-dropdownlist>
+</label>
