+++ examples-standalone/grid-charts-integration/src/app/components/common/select-chart-type/select-chart-type.component.ts
@@ -1,0 +1,15 @@
+import { Component, Input, Output, EventEmitter } from "@angular/core";
+
+@Component({
+    selector: "select-chart-type",
+    templateUrl: './select-chart-type.component.html'
+})
+export class SelectChartTypeComponent {
+    @Input() public data: Object[];
+    @Input() public chartName: string;
+    @Output() public valueChange = new EventEmitter<string>();
+
+    public onChange(chartName: string) {
+        this.valueChange.emit(chartName);
+    }
+}
