+++ examples-standalone/grid-charts-integration/src/app/components/common/select-series-btn-group/select-series-btn-group.component.html
@@ -1,0 +1,10 @@
+<label>Selected series:
+	<kendo-buttongroup>
+		<button #button *ngFor="let series of data" kendoButton
+			[toggleable]="true"
+			[selected]="isSelected(series)"
+			(selectedChange)="selectedChange($event, series)">
+			{{series.title}}
+		</button>
+	</kendo-buttongroup>
+</label>
