+++ examples-standalone/grid-charts-integration/src/app/components/common/window/window.component.html
@@ -1,0 +1,20 @@
+<stocks
+    *ngIf="isSimpleChart()"
+    [data]="data"
+    [chartConfiguration]="chartConfiguration"
+>
+</stocks>
+
+<pie-donut-stocks
+    *ngIf="isCircularChart()"
+    [data]="data"
+    [chartConfiguration]="chartConfiguration"
+>
+</pie-donut-stocks>
+
+<scatter-bubble-charts
+    *ngIf="isBubbleOrSeriesChart()"
+    [data]="data"
+    [chartConfiguration]="chartConfiguration"
+>
+</scatter-bubble-charts>
